curvetopia/whiteboard/image_processor.py
import numpy as np
import cv2
from scipy.optimize import curve_fit
from scipy.ndimage import rotate

def process_image(points):
    # Convert points to image
    image = points_to_image(points)
    
    # Apply image processing techniques
    symmetry = detect_symmetry(image)
    completed_image = complete_curves(image)
    fitted_points = fit_bezier_curve(points)
    
    # Combine results (this is a placeholder - you'll need to implement this based on your requirements)
    result = fitted_points  # For now, just return the fitted points
    
    return result


def points_to_image(points):
    # Create a blank image
    image = np.zeros((500, 500), dtype=np.uint8)
    
    # Draw points on the image
    for point in points:
        cv2.circle(image, tuple(point.astype(int)), 1, 255, -1)
    
    return image

def detect_symmetry(image, threshold=0.8, angle_step=1):
    # Convert to grayscale if not already
    if len(image.shape) > 2:
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
    else:
        gray = image.copy()

    # Binarize the image
    _, binary = cv2.threshold(gray, 127, 255, cv2.THRESH_BINARY)

    height, width = binary.shape
    best_score = 0
    best_symmetry = None

    # Check for vertical and horizontal symmetry
    for is_vertical in [True, False]:
        if is_vertical:
            range_to_check = width
        else:
            range_to_check = height
            binary = cv2.transpose(binary)

        for axis in range(1, range_to_check):
            left = binary[:, :axis]
            right = binary[:, axis:]
            right_flipped = cv2.flip(right, 1)

            # Pad the smaller side to match the larger side's width
            if left.shape[1] > right_flipped.shape[1]:
                right_flipped = np.pad(right_flipped, ((0, 0), (0, left.shape[1] - right_flipped.shape[1])), mode='constant')
            elif left.shape[1] < right_flipped.shape[1]:
                left = np.pad(left, ((0, 0), (0, right_flipped.shape[1] - left.shape[1])), mode='constant')

            # Compare left and flipped right sides
            match = np.sum(left == right_flipped) / (binary.shape[0] * max(left.shape[1], right_flipped.shape[1]))

            if match > best_score:
                best_score = match
                best_symmetry = ('vertical', axis) if is_vertical else ('horizontal', axis)

        if not is_vertical:
            binary = cv2.transpose(binary)

    # Check for rotational symmetry
    center = (width // 2, height // 2)
    for angle in range(1, 180, angle_step):
        rotated = rotate(binary, angle, reshape=False)
        match = np.sum(binary == rotated) / (height * width)

        if match > best_score:
            best_score = match
            best_symmetry = ('rotational', angle)

    if best_score >= threshold:
        return best_symmetry
    else:
        return None

# Example usage
def visualize_symmetry(image, symmetry):
    result = image.copy()
    height, width = image.shape[:2]
    
    if symmetry[0] == 'vertical':
        cv2.line(result, (symmetry[1], 0), (symmetry[1], height), (0, 255, 0), 2)
    elif symmetry[0] == 'horizontal':
        cv2.line(result, (0, symmetry[1]), (width, symmetry[1]), (0, 255, 0), 2)
    elif symmetry[0] == 'rotational':
        center = (width // 2, height // 2)
        cv2.circle(result, center, min(width, height) // 4, (0, 255, 0), 2)
        cv2.putText(result, f"{symmetry[1]}°", (center[0] + 10, center[1] + 10),
                    cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 1)
    
    return result

def complete_curves(image, max_gap=20):
    # Convert to grayscale if not already
    if len(image.shape) > 2:
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
    else:
        gray = image

    # Binarize the image
    _, binary = cv2.threshold(gray, 127, 255, cv2.THRESH_BINARY)

    # Find contours
    contours, _ = cv2.findContours(binary, cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)

    completed_image = np.zeros_like(binary)

    for contour in contours:
        # Check if contour is open
        if cv2.arcLength(contour, True) - cv2.arcLength(contour, False) > 1:
            # Find endpoints
            dists = np.sum(np.diff(contour[:, 0], axis=0)**2, axis=1)
            endpoints = np.argpartition(dists, -2)[-2:]

            start = tuple(contour[endpoints[0]][0])
            end = tuple(contour[endpoints[1]][0])

            # Check if endpoints are close enough
            if np.linalg.norm(np.array(start) - np.array(end)) <= max_gap:
                # Complete the curve
                cv2.line(completed_image, start, end, 255, 1)

        # Draw the original contour
        cv2.drawContours(completed_image, [contour], 0, 255, 1)

    return completed_image


def fit_bezier_curve(points):
    def cubic_bezier(t, p0, p1, p2, p3):
        return (1-t)**3 * p0 + 3*(1-t)**2*t * p1 + 3*(1-t)*t**2 * p2 + t**3 * p3

    def bezier_curve(t, p0x, p0y, p1x, p1y, p2x, p2y, p3x, p3y):
        x = cubic_bezier(t, p0x, p1x, p2x, p3x)
        y = cubic_bezier(t, p0y, p1y, p2y, p3y)
        return np.column_stack((x, y))

    t = np.linspace(0, 1, len(points))
    x, y = points[:, 0], points[:, 1]
    
    initial_guess = [x[0], y[0], x[1], y[1], x[-2], y[-2], x[-1], y[-1]]
    params, _ = curve_fit(bezier_curve, t, np.column_stack((x, y)), p0=initial_guess)
    
    fitted_points = bezier_curve(t, *params)
    return fitted_points