requirements.txt
Django==5.0.7
torch
sympy==1.13.1
pillow==10.4.0
pandas==2.2.2
opencv-python==4.10.0.84
numpy==2.0.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==20.1.0